+++ frontend/wailsjs/go/models.ts
@@ -558,6 +558,14 @@ export namespace ent {
 
 export namespace state {
 	
+	export enum BackupStatus {
+	    idle = "idle",
+	    waiting = "waiting",
+	    running = "running",
+	    completed = "completed",
+	    cancelled = "cancelled",
+	    failed = "failed",
+	}
 	export enum RepoStatus {
 	    idle = "idle",
 	    backingUp = "backingUp",
@@ -574,14 +582,6 @@ export namespace state {
 	    locked = "locked",
 	    unmount = "unmount",
 	    busy = "busy",
-	}
-	export enum BackupStatus {
-	    idle = "idle",
-	    waiting = "waiting",
-	    running = "running",
-	    completed = "completed",
-	    cancelled = "cancelled",
-	    failed = "failed",
 	}
 	export class BackupState {
 	    status: BackupStatus;
